Move tests for URL construction to test_Entrez.py
