+++ Tests/test_Entrez_parser.py
@@ -6,7 +6,6 @@
 """Testing code for Bio.Entrez parsers."""
 
 import unittest
-import warnings
 
 import os
 if os.name == 'java':
@@ -29,72 +28,9 @@ class GeneralTests(unittest.TestCase):
         handle.close()
         self.assertRaises(IOError, Entrez.read, handle)
 
-    def test_construct_cgi_ecitmatch(self):
-        citation = {
-            "journal_title": "proc natl acad sci u s a",
-            "year": "1991", "volume": "88", "first_page": "3248",
-            "author_name": "mann bj", "key": "citation_1"
-        }
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/ecitmatch.cgi'
-        variables = Entrez._update_ecitmatch_variables({'db': 'pubmed',
-                                                        'bdata': [citation]})
-        post = False
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=True, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertTrue("retmode=xml" in result_url, result_url)
-
-    def test_email_warning(self):
-        """Test issuing warning when user does not specify email address."""
-        Entrez.email = None
-
-        with warnings.catch_warnings(record=True) as w:
-            Entrez._construct_params(params=None)
-            self.assertEqual(len(w), 1)
-
-    def test_doing_get_request(self):
-        """By default we do GET requests to NCBI."""
-        params = Entrez._construct_params(None)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        post = Entrez._should_do_post_request(options, params)
-        self.assertFalse(post)
-
-    def test_doing_post_request(self):
-        """Test forcing POST request if 200+ UIDs are given."""
-        ids = [str(110000 + i) for i in range(202)]
-
-        params = Entrez._construct_params({'id': ids})
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        post = Entrez._should_do_post_request(options, params)
-        self.assertTrue(post)
-
-        params = Entrez._construct_params({'id': ",".join(ids)})
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        post = Entrez._should_do_post_request(options, params)
-        self.assertTrue(post)
-
-
-# This lets us set the email address to be sent to NCBI Entrez:
-Entrez.email = "biopython-dev@biopython.org"
-
-URL_HEAD = "http://eutils.ncbi.nlm.nih.gov/entrez/eutils/"
-URL_TOOL = "tool=biopython"
-URL_EMAIL = "email=biopython-dev%40biopython.org"
-
 
 class EInfoTest(unittest.TestCase):
     """Tests for parsing XML output returned by EInfo."""
-    def test_construct_cgi_einfo(self):
-        """Test constructed url for request to Entrez."""
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/einfo.fcgi'
-        params = Entrez._construct_params(params=None)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=False, options=options)
-        self.assertTrue(result_url.startswith(URL_HEAD + "einfo.fcgi?"), result_url)
-        self.assertTrue(URL_TOOL in result_url)
-        self.assertTrue(URL_EMAIL in result_url)
-
     def test_list(self):
         """Test parsing database list returned by EInfo."""
         # To create the XML file, use
@@ -1261,26 +1197,6 @@ class ESearchTest(unittest.TestCase):
 
 class EPostTest(unittest.TestCase):
     """Tests for parsing XML output returned by EPost."""
-    def test_construct_cgi_epost1(self):
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/epost.fcgi'
-        variables = {'db': 'nuccore', 'id': '186972394,160418'}
-        post = True
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertEqual(URL_HEAD + "epost.fcgi", result_url)
-
-    def test_construct_cgi_epost2(self):
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/epost.fcgi'
-        variables = {'db': 'nuccore', 'id': ["160418", "160351"]}
-        post = True
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertEqual(URL_HEAD + "epost.fcgi", result_url)
-
     # Don't know how to get an InvalidIdList in the XML returned by EPost;
     # unable to test if we are parsing it correctly.
     def test_epost(self):
@@ -1664,53 +1580,6 @@ class ESummaryTest(unittest.TestCase):
 
 class ELinkTest(unittest.TestCase):
     """Tests for parsing XML output returned by ELink."""
-    def test_construct_cgi_elink1(self):
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/elink.fcgi'
-        variables = {'cmd': 'neighbor_history', 'db': 'nucleotide',
-                     'dbfrom': 'protein', 'id': '22347800,48526535',
-                     'query_key': None, 'webenv': None}
-        post = False
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertTrue(result_url.startswith(URL_HEAD + "elink.fcgi?"), result_url)
-        self.assertTrue(URL_TOOL in result_url)
-        self.assertTrue(URL_EMAIL in result_url)
-        self.assertTrue("id=22347800%2C48526535" in result_url, result_url)
-
-    def test_construct_cgi_elink2(self):
-        """Commas: Link from protein to gene."""
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/elink.fcgi'
-        variables = {'db': 'gene', 'dbfrom': 'protein',
-                     'id': '15718680,157427902,119703751'}
-        post = False
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertTrue(result_url.startswith(URL_HEAD + "elink.fcgi"), result_url)
-        self.assertTrue(URL_TOOL in result_url)
-        self.assertTrue(URL_EMAIL in result_url)
-        self.assertTrue("id=15718680%2C157427902%2C119703751" in result_url, result_url)
-
-    def test_construct_cgi_elink3(self):
-        """Multiple ID entries: Find one-to-one links from protein to gene."""
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/elink.fcgi'
-        variables = {'db': 'gene', 'dbfrom': 'protein',
-                     'id': ["15718680", "157427902", "119703751"]}
-        post = False
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertTrue(result_url.startswith(URL_HEAD + "elink.fcgi"), result_url)
-        self.assertTrue(URL_TOOL in result_url)
-        self.assertTrue(URL_EMAIL in result_url)
-        self.assertTrue("id=15718680" in result_url, result_url)
-        self.assertTrue("id=157427902" in result_url, result_url)
-        self.assertTrue("id=119703751" in result_url, result_url)
-
     def test_pubmed1(self):
         """Test parsing pubmed links returned by ELink (first test)."""
         # Retrieve IDs from PubMed for PMID 9298984 to the PubMed database
@@ -3224,20 +3093,6 @@ class ESpellTest(unittest.TestCase):
 
 class EFetchTest(unittest.TestCase):
     """Tests for parsing XML output returned by EFetch."""
-    def test_construct_cgi_efetch(self):
-        cgi = 'http://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi'
-        variables = {'db': 'protein', 'id': '15718680,157427902,119703751',
-                     'retmode': 'xml'}
-        post = False
-
-        params = Entrez._construct_params(variables)
-        options = Entrez._encode_options(ecitmatch=False, params=params)
-        result_url = Entrez._construct_cgi(cgi, post=post, options=options)
-        self.assertTrue(result_url.startswith(URL_HEAD + "efetch.fcgi?"), result_url)
-        self.assertTrue(URL_TOOL in result_url)
-        self.assertTrue(URL_EMAIL in result_url)
-        self.assertTrue("id=15718680%2C157427902%2C119703751" in result_url, result_url)
-
     def test_pubmed1(self):
         """Test parsing XML returned by EFetch, PubMed database (first test)."""
         # In PubMed display PMIDs 12091962 and 9997 in xml retrieval mode
